requirements.txt
python-dotenv
ipykernel
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
pymongo[srv]>=4.0
certifi
mlflow
dagshub
fastapi
uvicorn
-e .

